♻️ refactor: child.ts 분리
